Added ValidateFilms.java and ValidateUsers.java files for validation purposes
