+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.models.Film;
+import ru.yandex.practicum.filmorate.validation.ValidateFilms;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -20,6 +21,7 @@ public class FilmController {
 
     private static final Logger LOG = LoggerFactory.getLogger(FilmController.class);
     private Map<Integer, Film> list = new HashMap<>();
+    private ValidateFilms validateFilms = new ValidateFilms();
 
     @GetMapping
     public List<Film> getAll() {
@@ -29,7 +31,8 @@ public class FilmController {
 
     @PostMapping
     public Film addFilm(@RequestBody Film film) {
-        if (validateName(film) && validateDuration(film) && validateDescription(film) && validateReleaseDate(film)) {
+        if (validateFilms.validateName(film) && validateFilms.validateDuration(film) &&
+                validateFilms.validateDescription(film) && validateFilms.validateReleaseDate(film)) {
             LOG.debug("Фильм прошёл валидацию и добавлен в базу");
             if (film.getId() == 0)
                 film.setId(1);
@@ -41,7 +44,8 @@ public class FilmController {
     @PutMapping
     public Film updateFilm(@RequestBody Film film) {
         Film newFilm = null;
-        if (validateName(film) && validateDuration(film) && validateDescription(film) && validateReleaseDate(film)) {
+        if (validateFilms.validateName(film) && validateFilms.validateDuration(film) &&
+                validateFilms.validateDescription(film) && validateFilms.validateReleaseDate(film)) {
             LOG.debug("Фильм прошёл валидацию при обновлении");
             if (list.containsKey(film.getId())) {
                 list.remove(film.getId());
@@ -59,52 +63,6 @@ public class FilmController {
             }
         }
         return newFilm;
-    }
-
-    public boolean validateName(Film film) {
-        if (film == null || film.getName() == null) {
-            LOG.error("Название фильма пустое!");
-            throw new ValidationException("Название фильма пустое!");
-        } else if (film.getName().length() == 0) {
-            LOG.error("Название фильма пустое!");
-            throw new ValidationException("Название фильма пустое!");
-        } else {
-            return true;
-        }
-    }
-
-    public boolean validateDescription(Film film) {
-        if (film == null || film.getDescription() == null) {
-            LOG.error("Описание фильма равно null");
-            throw new ValidationException("Описание фильма равно null");
-        } else if (film.getDescription().length() > 200) {
-            LOG.error("Описание фильма длиннее 200 знаков!");
-            throw new ValidationException("Описание фильма длиннее 200 знаков!");
-        } else {
-            return true;
-        }
-    }
-
-    public boolean validateReleaseDate(Film film) {
-        if (film == null || film.getReleaseDate() == null) {
-            LOG.error("Дата релиза фильма равна null");
-            throw new ValidationException("Дата релиза фильма равна null");
-        }
-        if (film.getReleaseDate().isBefore(LocalDate.parse("1895-12-28"))) {
-            LOG.error("Дата релиза фильма должна быть не раньше 28 декабря 1895 года!");
-            throw new ValidationException("Дата релиза фильма должна быть не раньше 28 декабря 1895 года!");
-        } else {
-            return true;
-        }
-    }
-
-    public boolean validateDuration(Film film) {
-        if (film.getDuration() <= 0) {
-            LOG.error("Продолжительность фильма должна быть положительной");
-            throw new ValidationException("Продолжительность фильма должна быть положительной");
-        } else {
-            return true;
-        }
     }
 
     public Map<Integer, Film> returnList() {
