+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.models.User;
+import ru.yandex.practicum.filmorate.validation.ValidateUsers;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -20,6 +21,7 @@ public class UserController {
 
     private static final Logger LOG = LoggerFactory.getLogger(UserController.class);
     private Map<Integer, User> list = new HashMap<>();
+    private ValidateUsers validateUsers = new ValidateUsers();
 
     @GetMapping
     public List<User> getAll() {
@@ -29,13 +31,14 @@ public class UserController {
 
     @PostMapping
     public User addUser(@RequestBody User user) {
-        if (validateBirthday(user) && validateEmail(user) && validateLogin(user)) {
+        if (validateUsers.validateBirthday(user) && validateUsers.validateEmail(user) &&
+                validateUsers.validateLogin(user)) {
             if (list.size() == 0) {
                 user.setId(1);
             } else {
                 user.setId(2);
             }
-            validateName(user);
+            validateUsers.validateName(user);
             list.put(user.getId(), user);
             LOG.debug("Пользователь успешно прошёл валидацию и добавлен в базу");
         }
@@ -46,8 +49,9 @@ public class UserController {
     public User updateUser(@RequestBody User user) {
         User newUser = null;
         if (list.containsKey(user.getId())) {
-            if (validateBirthday(user) && validateEmail(user) && validateLogin(user)) {
-                validateName(user);
+            if (validateUsers.validateBirthday(user) && validateUsers.validateEmail(user) &&
+                    validateUsers.validateLogin(user)) {
+                validateUsers.validateName(user);
                 newUser = new User();
                 newUser.setId(user.getId());
                 newUser.setName(user.getName());
@@ -64,55 +68,6 @@ public class UserController {
             throw new ValidationException("Пользователь \"" + user.getName() + "\" не найден в базе данных");
         }
         return newUser;
-    }
-
-    public boolean validateEmail(User user) {
-        if (user == null || user.getEmail() == null) {
-            LOG.error("Электронная почта пользователя равна null!");
-            throw new ValidationException("Электронная почта пользователя равна null!");
-        } else if (user.getEmail().isEmpty() || !user.getEmail().contains("@")) {
-            LOG.error("Указана неверная электронная почта пользователя!");
-            throw new ValidationException("Указана неверная электронная почта пользователя!");
-        } else {
-            return true;
-        }
-    }
-
-    public boolean validateLogin(User user) {
-        if (user == null || user.getLogin() == null) {
-            LOG.error("Логин пользователя равен null!");
-            throw new ValidationException("Логин пользователя равен null!");
-        }
-        if (user.getLogin().isBlank()) {
-            LOG.error("Указан неверный логин пользователя!");
-            throw new ValidationException("Указан неверный логин пользователя!");
-        } else {
-            return true;
-        }
-    }
-
-    public User validateName(User user) {
-        if (user == null) {
-            LOG.error("Пользователь равен null, валидация имени невозможна!");
-            throw new ValidationException("Пользователь равен null, валидация имени невозможна!");
-        } else if (user.getName() == null) {
-            user.setName(user.getLogin());
-        } else if (user.getName().isEmpty()) {
-            user.setName(user.getLogin());
-        }
-        return user;
-    }
-
-    public boolean validateBirthday(User user) {
-        if (user == null || user.getBirthday() == null) {
-            LOG.error("Дата рождения пользователя равна null!");
-            throw new ValidationException("Дата рождения пользователя равна null!");
-        } else if (user.getBirthday().isAfter(LocalDate.now())) {
-            LOG.error("Дата рождения пользователя не может быть в будущем!");
-            throw new ValidationException("Дата рождения пользователя не может быть в будущем!");
-        } else {
-            return true;
-        }
     }
 
     public Map<Integer, User> returnList() {
