+++ src/test/java/ru/yandex/practicum/filmorate/FilmorateApplicationFilmsTests.java
@@ -6,6 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.models.Film;
 import ru.yandex.practicum.filmorate.controllers.FilmController;
+import ru.yandex.practicum.filmorate.validation.ValidateFilms;
 
 import java.time.LocalDate;
 
@@ -17,10 +18,12 @@ class FilmorateApplicationFilmsTests {
     private FilmController filmController;
     private Film film1;
     private Film film2;
+    private ValidateFilms validateFilms;
 
     @BeforeEach
     void init() {
         filmController = new FilmController();
+        validateFilms = new ValidateFilms();
 
         film1 = new Film();
         film1.setId(1);
@@ -50,7 +53,7 @@ class FilmorateApplicationFilmsTests {
                 "Описание фильма длиннее 200 знаков!Описание фильма длиннее 200 знаков!" +
                 "Описание фильма длиннее 200 знаков!Описание фильма длиннее 200 знаков!" +
                 "Описание фильма длиннее 200 знаков!Описание фильма длиннее 200 знаков!");
-        Exception exception = assertThrows(ValidationException.class, () -> filmController.validateDescription(film));
+        Exception exception = assertThrows(ValidationException.class, () -> validateFilms.validateDescription(film));
         assertEquals("Описание фильма длиннее 200 знаков!", exception.getMessage());
     }
 
@@ -58,14 +61,14 @@ class FilmorateApplicationFilmsTests {
     void testFilmDurationValidation() {
         final Film film = new Film();
         film.setDuration(-1);
-        Exception exception = assertThrows(ValidationException.class, () -> filmController.validateDuration(film));
+        Exception exception = assertThrows(ValidationException.class, () -> validateFilms.validateDuration(film));
         assertEquals("Продолжительность фильма должна быть положительной", exception.getMessage());
     }
 
     @Test
     void testFilmNameValidation() {
         final Film film = new Film();
-        Exception exception = assertThrows(ValidationException.class, () -> filmController.validateName(film));
+        Exception exception = assertThrows(ValidationException.class, () -> validateFilms.validateName(film));
         assertEquals("Название фильма пустое!", exception.getMessage());
     }
 
@@ -73,7 +76,7 @@ class FilmorateApplicationFilmsTests {
     void testFilmDateValidation() {
         final Film film = new Film();
         film.setReleaseDate(LocalDate.parse("1895-12-27"));
-        Exception exception = assertThrows(ValidationException.class, () -> filmController.validateReleaseDate(film));
+        Exception exception = assertThrows(ValidationException.class, () -> validateFilms.validateReleaseDate(film));
         assertEquals("Дата релиза фильма должна быть не раньше 28 декабря 1895 года!", exception.getMessage());
     }
 
@@ -81,7 +84,7 @@ class FilmorateApplicationFilmsTests {
     void testFilmDateValidationBorder() {
         final Film film = new Film();
         film.setReleaseDate(LocalDate.parse("1895-12-28"));
-        assertTrue(filmController.validateReleaseDate(film));
+        assertTrue(validateFilms.validateReleaseDate(film));
     }
 
     @Test
